refactor(package): 책임에 맞게 패키지 구조 리펙터링
