+++ src/main/java/lotto/domain/constant/LottoConstant.java
@@ -1,9 +1,9 @@
-package lotto.domain.policy;
+package lotto.domain.constant;
 
-public class LottoPolicy {
+public class LottoConstant {
     public static final int LOTTO_MIN_NUMBER = 1;
     public static final int LOTTO_MAX_NUMBER = 45;
     public static final int LOTTO_PRICE = 1000;
 
-    private LottoPolicy() {}
+    private LottoConstant() {}
 }
